+++ src/commands/CL08ledSetTimeOnOff.h
@@ -1,8 +1,8 @@
 void cL08ledSetTimeOnOff(){
     // Set On Off time to Led
-    size_t lednr = commandParts.arg1.toInt();
-    size_t tOn = commandParts.arg2.toInt();  //aparently toInt also converts to long
-    size_t tOff = commandParts.arg3.toInt();
+    uint8_t lednr = commandParts.arg1.toInt();
+    uint32_t tOn = commandParts.arg2.toInt();  //aparently toInt also converts to long
+    uint32_t tOff = commandParts.arg3.toInt();
     Println("L08 Set led " + commandParts.arg1 + " timeOn= " + commandParts.arg2 + ", timeOff= " + commandParts.arg3);
     if (lednr >= 0 && lednr < LED_NR_ITEMS) {
         if ((tOn >= 0) && (tOff >= 0)) {
