+++ src/commands/CL20ledSaveSetting.h
@@ -1,6 +1,6 @@
 void cL20ledSaveSetting(){
     // Save settings for a specific Led
-    size_t lednr = commandParts.arg1.toInt();
+    uint8_t lednr = commandParts.arg1.toInt();
     String animatie = commandParts.arg2;
 
     if (lednr >= 0 && lednr < LED_NR_ITEMS) {
