+++ src/commands/cA01applyAnim.h
@@ -1,6 +1,6 @@
 void cA01applyAnim(){
   // Apply animation
-  size_t anim = commandParts.arg1.toInt();
+  uint8_t anim = commandParts.arg1.toInt();
   if (anim < 0 || anim > HUIS_NR_ANIM) {
     Println(F("Invalid animation number"));
   }
