+++ src/commands/cR09rgbEffect.h
@@ -1,6 +1,6 @@
 void cR09rgbEffect(){
     // Set JPEG for rgb
-    size_t rgbnr = commandParts.arg1.toInt();
+    uint8_t rgbnr = commandParts.arg1.toInt();
     String name = commandParts.arg2;
     if (rgbnr >= 0 && rgbnr < RGB_NR_ITEMS) {
         rgbactive[rgbnr].desc = name;
