+++ src/rgb/rgbAnimApply.h
@@ -1,8 +1,8 @@
 #pragma once
 String rgbAnimApply() {
   // Apply RGB settings for given animatie
-  size_t rgbnr = rgbRequest.rgbnr;
-  size_t animnr = rgbRequest.animnr;
+  uint8_t rgbnr = rgbRequest.rgbnr;
+  uint8_t animnr = rgbRequest.animnr;
   if (huisSetting.animdesc[animnr] == NULL) {
     Println("rgbAnimApply: animatie niet gevonden nr " + String(animnr));
   }
