+++ src/rgb/rgb_uit.h
@@ -1,6 +1,6 @@
 #pragma once
-void rgb_uit(size_t rgbnr) {
-  for (size_t i = rgbactive[rgbnr].startrgb; i <= rgbactive[rgbnr].stoprgb; i++)
+void rgb_uit(uint8_t rgbnr) {
+  for (uint8_t i = rgbactive[rgbnr].startrgb; i <= rgbactive[rgbnr].stoprgb; i++)
   {
     rgbstrip[i] = CRGB::Black;
   }
